+++ Chrysalis/nbproject/private/private.properties
@@ -1,1 +1,2 @@
 compile.on.save=true
+user.properties.file=C:\\Users\\James\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
