+++ Chrysalis/nbproject/project.properties
@@ -26,6 +26,8 @@ dist.archive.excludes=
 dist.dir=dist
 dist.jar=${dist.dir}/Chrysalis.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+excludes=
+includes=**
 jar.compress=false
 javac.classpath=
 # Space-separated list of extra javac options
